Add icon automation
