+++ package.json
@@ -11,7 +11,8 @@
     "start": "gatsby serve",
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\"",
     "deploy": "gatsby build --prefix-paths && gh-pages -d public",
-    "convert": "for i in content/**/*.rst ; do python3 scripts/pre_pandoc_script.py ${i%}; echo \"$i\" && pandoc --wrap=none $i -f rst -t gfm -o ${i%.*}.mdx ; done ; python3 scripts/post_pandoc_script.py"
+    "convert": "for i in content/**/*.rst ; do python3 scripts/pre_pandoc_script.py ${i%}; echo \"$i\" && pandoc --wrap=none $i -f rst -t gfm -o ${i%.*}.mdx ; done ; python3 scripts/post_pandoc_script.py",
+    "update-icons": "git submodule update --init --remote && node scripts/createIconTypes.js && node scripts/createIconNames.js"
   },
   "dependencies": {
     "@mdx-js/mdx": "^1.6.1",
