+++ src/components/icon/iconNames.js
@@ -1,0 +1,66 @@
+export default {
+ ANNOUNCE: "announce",
+ ARROW_LEFT: "arrowleft",
+ ARROW_RIGHT: "arrowright",
+ BRAIN_CIRCUIT: "braincircuit",
+ BUSINESSMAN: "businessman",
+ BUSINESSWOMAN: "businesswoman",
+ CASE_STUDY: "casestudy",
+ CHECKMARK: "checkmark",
+ CHEVRON_DOWN: "chevrondown",
+ CHEVRON_RIGHT: "chevronright",
+ CLOSE: "close",
+ CLOUD_DB: "clouddb",
+ CLOUD_DBA: "clouddba",
+ CODE_WRITING: "codewriting",
+ CONNECT: "connect",
+ CONVERT: "convert",
+ DATABASE_ADMIN: "databaseadmin",
+ DELIVER_LOVE: "deliverlove",
+ DEVELOPER: "developer",
+ DOCS: "docs",
+ DOTTED_BOX: "dottedbox",
+ EARTH: "earth",
+ EASY: "easy",
+ EDB_BADGE: "edbbadge",
+ EDB_BART: "edbbart",
+ EDB_EFM: "edbefm",
+ EDB_EPAS: "edbepas",
+ EDB_MIGRATION: "edbmigration",
+ EDB_PEM: "edbpem",
+ EDB_REPLICATION: "edbreplication",
+ ENERGY: "energy",
+ EXPORT: "export",
+ FILE: "file",
+ HAMBURGER: "hamburger",
+ HANDSHAKE: "handshake",
+ HIGH_AVAILABILITY: "highavailability",
+ IDEA_SHARING: "ideasharing",
+ IMPROVE: "improve",
+ INSTALL: "install",
+ KNIGHT: "knight",
+ KUBERNETES: "kubernetes",
+ LAPTOP_CONFIG: "laptopconfig",
+ LEARNING: "learning",
+ LINKEDIN: "linkedin",
+ MISSION: "mission",
+ NETWORK: "network",
+ NETWORK2: "network2",
+ NEWS: "news",
+ PARTNER: "partner",
+ POSTGRES_SUPPORT: "postgressupport",
+ POSTGRESQL: "postgresql",
+ PRESENTATION: "presentation",
+ PROCESS_IMPROVEMENT: "processimprovement",
+ REMOTE_DBA: "remotedba",
+ ROCKET: "rocket",
+ SEARCH: "search",
+ STACK: "stack",
+ SUPPORT: "support",
+ TEN_YEARS: "tenyears",
+ TICKET: "ticket",
+ TOOLS: "tools",
+ TUTORIAL: "tutorial",
+ VENN: "venn",
+ WEBINAR: "webinar",
+}
