+++ src/components/layout.js
@@ -3,7 +3,7 @@ import { Helmet } from 'react-helmet';
 import useSiteMetadata from '../hooks/use-sitemetadata';
 import TextBalancer from '../components/text-balancer';
 import { MDXProvider } from '@mdx-js/react';
-import Icon from '../components/icon';
+import Icon from '../components/icon/';
 
 import '../styles/index.scss';
