+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container } from 'react-bootstrap';
 import { indexLinkList } from '../constants/index-link-list';
-import Icon, { iconNames } from '../components/icon';
+import Icon, { iconNames } from '../components/icon/';
 import {
   Footer,
   IndexLinks,
